Attachments: move GDrive/Web sources
